fix: Change expected rank type for ins_sort

+++ src/insertion_sort.c
@@ -1,18 +1,18 @@
-/* ************************************************************************** */
+/******************************************************************************/
 /*                                                                            */
 /*                                                        :::      ::::::::   */
 /*   insertion_sort.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: abisani <abisani@student.42.fr>            +#+  +:+       +#+        */
+/*   By: abisiani <abisiani@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/04 10:26:56 by abisani           #+#    #+#             */
-/*   Updated: 2025/11/04 11:15:27 by abisani          ###   ########.fr       */
+/*   Updated: 2025/11/06 13:00:10 by abisiani         ###   ########.fr       */
 /*                                                                            */
-/* ************************************************************************** */
+/******************************************************************************/
 
 #include "push_swap.h"
 
-static void	find_next_rank(t_stacks *stacks, size_t next_rank)
+static void	find_next_rank(t_stacks *stacks, int next_rank)
 {
 	while (stacks->a->rank != next_rank)
 		rotate(&(stacks->a), stacks->a);
